Move threadState type to debugPanel
